implementation checkpoint

+++ SistemInventarisLabApp/src/main/java/com/example/sisteminventarislab/web/controller/BarangController.java
@@ -3,6 +3,9 @@ package com.example.sisteminventarislab.web.controller;
 import com.example.sisteminventarislab.entity.Barang;
 import com.example.sisteminventarislab.entity.Response;
 import com.example.sisteminventarislab.entity.helper.ResponseHelper;
+import com.example.sisteminventarislab.exception.CustomException;
+import com.example.sisteminventarislab.exception.ErrorCode;
+import com.example.sisteminventarislab.service.AccessTokenService;
 import com.example.sisteminventarislab.service.BarangService;
 import com.example.sisteminventarislab.web.model.Request.CreateBarangWebRequest;
 import com.example.sisteminventarislab.web.model.Request.UpdateBarangWebRequest;
@@ -11,6 +14,8 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
+import org.springframework.util.ObjectUtils;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,14 +30,18 @@ import javax.validation.Valid;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotEmpty;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Api
 @RestController
-@RequestMapping(path = "/api/barang", produces = MediaType.APPLICATION_JSON_VALUE)
+@RequestMapping(path = "/api/barang",
+    produces = MediaType.APPLICATION_JSON_VALUE)
 @RequiredArgsConstructor
+@CrossOrigin
 public class BarangController {
 
   private final BarangService barangService;
+  private final AccessTokenService accessTokenService;
 
   /**
    * Method createBarang merupakan method untuk membuat barang baru
@@ -44,7 +53,11 @@ public class BarangController {
    */
   @ApiOperation("create new Barang")
   @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
-  public Response<CreateBarangWebResponse> createBarang(@RequestBody @Valid CreateBarangWebRequest request) {
+  public Response<CreateBarangWebResponse> createBarang(@RequestBody @Valid CreateBarangWebRequest request,
+      @RequestParam String token) {
+    if (!accessTokenService.doExist(token))
+      throw new CustomException(ErrorCode.UNAUTHORIZED);
+
     return ResponseHelper.ok(barangService.createBarang(request));
   }
 
@@ -71,7 +84,11 @@ public class BarangController {
   @GetMapping
   public Response<List<Barang>> getBarangListPaged(@RequestParam @Valid @NotEmpty(message = "Page tidak boleh kosong!") @Min(value = 1,
       message = "Page tidak boleh bernilai < 1!") Integer page) {
-    return ResponseHelper.ok(barangService.getAllBarangPaged(page));
+    return ResponseHelper.ok(barangService.getAllBarangPaged(page)
+        .stream()
+        .filter(barang -> ObjectUtils.isEmpty(barang.getIdPeminjam()))
+        .collect(
+            Collectors.toList()));
   }
 
   /**
@@ -92,7 +109,10 @@ public class BarangController {
       consumes = MediaType.APPLICATION_JSON_VALUE,
       path = "/{id}")
   public Response<Barang> updateBarang(@PathVariable String id,
-      @RequestBody @Valid UpdateBarangWebRequest request) {
+      @RequestBody @Valid UpdateBarangWebRequest request, @RequestParam String token) {
+    if (!accessTokenService.doExist(token))
+      throw new CustomException(ErrorCode.UNAUTHORIZED);
+
     return ResponseHelper.ok(barangService.updateBarang(id, request));
   }
 
@@ -109,7 +129,10 @@ public class BarangController {
    */
   @ApiOperation("delete Barang by id")
   @DeleteMapping(path = "/{id}")
-  public Response<Boolean> deleteBarang(@PathVariable String id) {
+  public Response<Boolean> deleteBarang(@PathVariable String id, @RequestParam String token) {
+    if (!accessTokenService.doExist(token))
+      throw new CustomException(ErrorCode.UNAUTHORIZED);
+
     return ResponseHelper.ok(barangService.deleteBarang(id));
   }
 }
